pyproject.toml
[project]
name = "reframebot"
version = "0.1.0"
description = "CBT-oriented chatbot for university students with academic stress"
requires-python = ">=3.11"

# Runtime — what you need to run `python app.py`
dependencies = [
    "fastapi>=0.121",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "torch>=2.0",
    "transformers>=4.40",
    "peft>=0.10",
    "bitsandbytes>=0.43",
    "accelerate>=0.30",
    "sentence-transformers>=3.0",
    "chromadb>=0.5",
    "numpy>=1.26",
]

[project.optional-dependencies]
# pip install -e ".[scripts]"  — evaluation + push scripts
scripts = [
    "datasets>=2.0",
    "scikit-learn>=1.4",
    "bert-score>=0.3",
    "huggingface-hub>=0.20",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
]
# pip install -e ".[train]"  — full training pipeline (includes scripts)
train = [
    "reframebot[scripts]",
    "trl>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reframebot"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
